provide timeInLoop and timeInLoopAsync from _internal
